pyproject.toml
[tool.poetry]
name = "crm-management"
version = "0.1.0"
description = ""
authors = ["jorge <jorge.olivares@somedomain.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
click = "^8.1.7"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
